+++ src/pczt/tx_extractor.rs
@@ -1,3 +1,5 @@
+use core::fmt;
+
 use nonempty::NonEmpty;
 use rand::{CryptoRng, RngCore};
 
@@ -115,6 +117,30 @@ pub enum TxExtractorError {
     /// The value sum does not fit into a `valueBalance`.
     ValueSumOutOfRange,
 }
+
+impl fmt::Display for TxExtractorError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            TxExtractorError::MissingBindingSignatureSigningKey => {
+                write!(f, "`bsk` must be set for the Transaction Extractor role")
+            }
+            TxExtractorError::MissingProof => write!(
+                f,
+                "Orchard `zkproof` must be set for the Transaction Extractor role"
+            ),
+            TxExtractorError::MissingSpendAuthSig => write!(
+                f,
+                "`spend_auth_sig` fields must all be set for the Transaction Extractor role"
+            ),
+            TxExtractorError::ValueSumOutOfRange => {
+                write!(f, "value sum does not fit into a `valueBalance`")
+            }
+        }
+    }
+}
+
+#[cfg(feature = "std")]
+impl std::error::Error for TxExtractorError {}
 
 /// Authorizing data for a bundle of actions that is just missing a binding signature.
 #[derive(Debug)]